feat: add shared navigation & footer to all public pages via route group
Public pages (/resume, /blog, /blog/[slug]) were dead ends with no navbar
or footer. This uses a (public) route group layout to share Navigation,
Footer, CustomCursor, and ScrollProgress across all public pages while
keeping /login and /admin/* unaffected.

- Navigation: logo links home, active state for current pathname
- Footer: uses Link for path-based nav, anchor for hash links
