+++ src/app/(public)/blog/[slug]/page.tsx
